+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Added `.jar` files to enable pip install from GitHub
 ### Changed 
+- Removed `.mean()` for GCI losses in BoxSquaredELModule
 ### Deprecated
 ### Removed
 ### Fixed